separating lx200generic and lx200telescope
